scripts/requirements.txt
discord.py
httpx
python-dotenv
